+++ src/main/java/com/example/gak/domain/chatmessage/entity/ChatMessage.java
@@ -2,7 +2,7 @@ package com.example.gak.domain.chatmessage.entity;
 
 import com.example.gak.domain.common.entity.BaseEntity;
 import com.example.gak.domain.member.entity.Member;
-import com.example.gak.domain.sessionroom.entity.SessionRoom;
+import com.example.gak.domain.session.entity.SessionRoom;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
